Gross IRPEF
- Formula to calculate gross IRPEF and relative test.
